Moved shared code into another module
