+++ tasks/publish-matrix.ts
@@ -1,19 +1,6 @@
-import { call, main, type Operation } from "effection";
+import { call, main } from "effection";
 import { x } from "../tinyexec/mod.ts";
-import { z } from "npm:zod@3.23.8";
-import { resolve } from "jsr:@std/path@^1.0.6";
-
-export const DenoJson = z.object({
-  name: z.string(),
-  version: z.string(),
-  exports: z.union([z.record(z.string()), z.string()]),
-  license: z.string(),
-});
-
-type PackageConfig = {
-  workspace: string;
-  workspacePath: string;
-} & z.infer<typeof DenoJson>;
+import { readPackages } from "./lib/read-packages.ts";
 
 await main(function* () {
   let packages = yield* readPackages();
@@ -61,32 +48,3 @@ await main(function* () {
     );
   }
 });
-
-function* readPackages(): Operation<PackageConfig[]> {
-  const root = yield* call(() =>
-    import("../deno.json", {
-      with: { type: "json" },
-    })
-  );
-
-  console.log(`Found ${root.default.workspace.join(", ")}`);
-
-  const configs: PackageConfig[] = [];
-  for (let workspace of root.default.workspace) {
-    const workspacePath = resolve(Deno.cwd(), workspace);
-
-    const config = yield* call(() =>
-      Deno.readTextFile(`${workspacePath}/deno.json`)
-    );
-
-    const denoJson = DenoJson.parse(JSON.parse(config));
-
-    configs.push({
-      ...denoJson,
-      workspace: workspace.replace("./", ""),
-      workspacePath,
-    });
-  }
-
-  return configs;
-}
